+++ hive/readme/store_objects/hive_adapters.dart
@@ -1,7 +1,7 @@
 import 'package:hive_ce/hive.dart';
 import 'person.dart';
 
-part 'hive_adapters.hive.dart';
+part 'hive_adapters.g.dart';
 
 @GenerateAdapters([AdapterSpec<Person>()])
 // Annotations must be on some element