+++ hive_generator/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 NOTE: This is a breaking change. See the [migration guide](MIGRATION.md) for more information.
 
-- **BREAKING**: Moves `hive_registrar.g.dart` to `hive/hive_registrar.hive.dart`
+- **BREAKING**: Moves `hive_registrar.g.dart` to `hive/hive_registrar.g.dart`
 
 ## 1.7.0